--- packages/chains/vite.common.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig } from 'vite';
-import tsconfigPaths from 'vite-tsconfig-paths';
-
-export default defineConfig({
-  plugins: [tsconfigPaths()],
-});
