+++ packages/chains/vite.config.ts
@@ -1,24 +1,20 @@
-import { defineConfig, mergeConfig } from 'vite';
+import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
+import tsconfigPaths from 'vite-tsconfig-paths';
 
-// configs
-import commonConfig from './vite.common.config';
-
-export default mergeConfig(
-  commonConfig,
-  defineConfig({
-    build: {
-      lib: {
-        entry: 'src/index.ts',
-        formats: ['es'],
-        fileName: 'index',
-      },
-      outDir: 'dist',
+export default defineConfig({
+  build: {
+    lib: {
+      entry: 'src/index.ts',
+      formats: ['es'],
+      fileName: 'index',
     },
-    plugins: [
-      dts({
-        tsconfigPath: 'tsconfig.build.json',
-      }),
-    ],
-  })
-);
+    outDir: 'dist',
+  },
+  plugins: [
+    dts({
+      tsconfigPath: 'tsconfig.build.json',
+    }),
+    tsconfigPaths(),
+  ],
+});
